added plus to pass on validations and a bit more refactoring

+++ checker_bonus/utils/libft/ft_atoi_bonus.c
@@ -12,9 +12,9 @@
 
 #include "../../checker_bonus.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	ret;
+	long	ret;
 	int	sign;
 	int	i;
